+++ packages/bundlekit-bundler-rollup/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @bundlekit/bundler-rollup
 
+## 0.0.17
+
+### Patch Changes
+
+- 71acfd9: 增加 angular 得模版
+- c00d434: 增加 svelte 模版支持
+- Updated dependencies [71acfd9]
+- Updated dependencies [c00d434]
+  - @bundlekit/shared-utils@0.0.10
+
 ## 0.0.16
 
 ### Patch Changes