+++ packages/bundlekit-plugin-vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bundlekit/plugin-vue
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [71acfd9]
+- Updated dependencies [c00d434]
+  - @bundlekit/shared-utils@0.0.10
+
 ## 0.0.12
 
 ### Patch Changes